requirements.txt
python-chess>=1.999
click>=8.0
rich>=13.0
